refactor: cp
